+++ CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Noteworthy
 
-  -
+  - added `discriminator` to `IEntityRepositoryOptions` which should allow limiting number of sequentially scanned 
+    entries
 
 ## [4.9.0] - 2021-11-10